// src/app/task-detail/modules/sidebar-right/modules/widget-rent-manager/modules/rent-manager-lease-renewal/enums/rent-manager-lease-renewal.enum.ts
export enum EWeekly {
  MONDAY = 1,
  TUESDAY = 2,
  WEDNESDAY = 3,
  THURSDAY = 4,
  FRIDAY = 5,
  SATURDAY = 6,
  SUNDAY = 7
}

export enum ERecurringChargeMode {
  VIEW = 'VIEW',
  EDIT = 'EDIT',
  ADD = 'ADD'
}

export enum EFieldType {
  YESNO = 'YesNo',
  DATE = 'Date',
  NUMERIC = 'Numeric',
  CHECKED_LIST = 'Checkedlist',
  ENCRYPTED_TEXT = 'EncryptedText',
  DROPDOWN = 'Dropdown',
  FILE = 'File',
  HYPERLINK = 'Hyperlink',
  IMAGE = 'Image'
}

export enum EAttachment {
  FILE = 'file',
  PHOTO = 'photo'
}